requirements.txt
asgiref==3.8.1
Django==5.1
django-appconf==1.0.6
django-compressor==4.5.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pytz==2024.1
PyYAML==6.0.2
rcssmin==1.1.2
rjsmin==1.2.2
setuptools==72.2.0
sqlparse==0.5.1
uritemplate==4.1.1
whitenoise==6.7.0
